requirements.txt
tensorflow==2.13.0
protobuf<4
streamlit
matplotlib
seaborn
pillow
